eopkg_enumerator: Add skeletal code for an EopkgEnumerator
The current code is very janky so we'll need to define a stable representation
of an `eopkg` package for serialisation and storage before we can even add a
database cache.

We also want to ensure tight control of python invocations, so we'll wrap the
enumeration up in a RAII range.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>